requirements.txt
pywin32
psutils
# ttkbootstrap
customtkinter